docs + test + refactor : 📝 Test & write API Blueprint docs
