+++ src/app/store/commands/create-store.command.ts
@@ -1,7 +1,7 @@
 import {ICommand} from '@nestjs/cqrs';
 import {Store} from '@store';
 
-export class InsertStoreCommand extends Store implements ICommand {
+export class CreateStoreCommand extends Store implements ICommand {
 	constructor(args: any) {
 		super(args);
 	}
