+++ src/app/store/repository/store.repository.ts
@@ -1,6 +1,10 @@
-import {EntityRepository, Repository} from 'typeorm';
+import {EntityRepository, Repository, Transaction, TransactionManager} from 'typeorm';
 import {Store} from '@store';
 
 @EntityRepository(Store)
 export class StoreRepository extends Repository<Store> {
+	@Transaction()
+	public async saveStore(store: Store, @TransactionManager() storeRepository: Repository<Store>) {
+		return storeRepository.save(store);
+	}
 }
